Converted reporter to a class, Correctly fetching config options
